+++ test/Parity Tests/Aggregate Tests/Aggregate Tests.js
@@ -33,6 +33,7 @@ module.exports = function ( Driver )
 		require( './test-suite/Type Operator Tests.js' )( Driver );
 		require( './test-suite/Data Size Operator Tests.js' )( Driver );
 		require( './test-suite/Date Operator Tests.js' )( Driver );
+		require( './test-suite/Set Operator Tests.js' )( Driver );
 		require( './test-suite/Stage and Accumulator Tests.js' )( Driver );
 		require( './test-suite/Expression Rejection Tests.js' )( Driver );
 	} );